hotels | index | error update

+++ controllers/hotels.js
@@ -1,17 +1,17 @@
 import Hotel from "../models/Hotel.js";
+import { createError } from "../utils/error.js";
 
-export const createHotel = async (req, res) => {
+export const createHotel = async (req, res, next) => {
   const hotel = Hotel(req.body);
   try {
     const savedHotel = await hotel.save();
     res.status(201).json(savedHotel);
   } catch (error) {
-    console.log(error);
-    res.status(500).json(error);
+    next(createError(error));
   }
 };
 
-export const updateHotel = async (req, res) => {
+export const updateHotel = async (req, res, next) => {
   try {
     const updatedHotel = await Hotel.findByIdAndUpdate(
       req.params.id,
@@ -22,37 +22,44 @@ export const updateHotel = async (req, res) => {
     );
     res.status(201).json(updatedHotel);
   } catch (error) {
-    console.log(error);
-    res.status(500).json(error);
+    next(createError(error));
   }
 };
-export const deleteHotel = async (req, res) => {
+export const deleteHotel = async (req, res, next) => {
   try {
+    const hotel = await Hotel.findByIdAndDelete(req.params.id);
+
+    if (!hotel) {
+      next(
+        createError({
+          status: 404,
+          message: "The hotel you are trying to delete doesnt exist",
+        })
+      );
+      return;
+    }
+
     await Hotel.findByIdAndDelete(req.params.id);
     res.status(200).json("Hotel Has Been Deleted");
   } catch (error) {
-    console.log(error);
-    res.status(500).json(error);
+    next(createError(error));
   }
 };
 
-export const getHotel = async (req, res) => {
+export const getHotel = async (req, res, next) => {
   try {
     const hotel = await Hotel.findById(req.params.id);
     res.status(200).json(hotel);
   } catch (error) {
-    console.log(error);
-    res.status(500).json(error);
+    next(createError(error));
   }
 };
 
-
-export const getAllHotels = async (req, res) => {
+export const getAllHotels = async (req, res, next) => {
   try {
     const hotels = await Hotel.find();
     res.status(200).json(hotels);
   } catch (error) {
-    console.log(error);
-    res.status(500).json(error);
+    next(createError(error));
   }
 };
